Break out convolution into its own module
