+++ src/layer/convolutional.rs
@@ -1,6 +1,6 @@
 use super::SerializedLayer;
 use crate::tensor::helper::TensorShape;
-use crate::tensor::Tensor;
+use crate::tensor::{self, Tensor};
 use crate::{layer::Layer, network::NetworkRunState};
 use std::collections::HashMap;
 
@@ -171,7 +171,13 @@ impl Layer for ConvolutionalLayer {
     fn forward(self: &mut Self, input: &Tensor, output: &mut Tensor) -> Result<(), &'static str> {
         // Convolution for every output channel
         for i in 0..self.output_shape.dk {
-            Tensor::convolution(input, &self.kernels[i as usize], self.stride, output, i);
+            tensor::operations::convolution::convolution(
+                input,
+                &self.kernels[i as usize],
+                self.stride,
+                output,
+                i,
+            );
         }
         return Ok(());
     }
